refactor: Binding Conversions samples, packages renamed
