+++ app/build.gradle
@@ -38,10 +38,7 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation files('libs/YouTubeAndroidPlayerApi.jar')
 
-//    def lifecycle_version = "2.1.0-alpha04"
-//    implementation "androidx.lifecycle:lifecycle-compiler:$lifecycle_version"
-//    annotationProcessor "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
-
+    // ARCHITECTURE COMPONENTS DEPENDENCIES
     def lifecycle_version = "1.1.1"
     implementation "android.arch.lifecycle:extensions:$lifecycle_version"
     annotationProcessor "android.arch.lifecycle:compiler:$lifecycle_version"
@@ -61,6 +58,8 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.5.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.5.0'
     implementation 'com.github.bumptech.glide:glide:4.8.0'
+
+    //
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
