+++ app/src/main/java/com/udacity/popularmovies/service/MovieDataService.java
@@ -2,8 +2,8 @@ package com.udacity.popularmovies.service;
 
 
 import com.udacity.popularmovies.config.ApiConfig.UrlParamKey;
-import com.udacity.popularmovies.moviedetails.models.MovieDetailsResponse;
-import com.udacity.popularmovies.moviesfeed.models.MoviesResponse;
+import com.udacity.popularmovies.moviedetails.model.MovieDetailsResponse;
+import com.udacity.popularmovies.moviesfeed.model.MoviesResponse;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
