+++ app/src/main/res/values/strings.xml
@@ -101,4 +101,5 @@
     <string name="no_home_page_available">This movie has no homepage available.</string>
 
     <string name="appbar_scrolling_behavior" translatable="false">com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior</string>
+    <string name="not_available">Not available</string>
 </resources>
